refactor: consolidate service packages and update references
Move individual service packages into a single 'service' package and update all references to use the new structure. This includes renaming and moving files for user, community, post, and comment services, as well as updating the ServiceManager and worker implementations to align with the new package structure.
