+++ internal/service/posts.go
@@ -1,4 +1,4 @@
-package posts
+package service
 
 import (
 	"sublinks/sublinks-federation/internal/db"
